server/requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
keystone-engine==0.9.2
MarkupSafe==2.0.1
python-dotenv==0.17.1
six==1.16.0
unicorn==1.0.3
Werkzeug==2.0.1
